requirements.txt
line-bot-sdk
flask
requests
hanziconv
beautifulsoup4